# docs/api.yaml
openapi: '3.0.2'
info:
  title: Reactions Storage Service
  description: Service responsible for configuring and managing user reactions
  version: '1.0'
tags:
  - name: UserReactions
    description: |
      Operations related to user reactions.
      Should be used by other services
  - name: ReactionsRepo
    description: |
      Operations related to reactions repo (reactions configuration).
      Should be used by developers
paths:
  /reactions:
    get:
      description: Retrieve user reactions info for given entity
      parameters:
        - in: query
          name: namespace_id
          required: true
          schema:
            type: string
        - in: query
          name: entity_id
          required: true
          schema:
            type: string
        - in: query
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/ReactionsGETResponse200'
    post:
      description: Add single user reaction
      parameters:
        - in: query
          name: force
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'definitions/requests.yaml#/components/schemas/ReactionsPOSTRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/ReactionsPOSTResponse200'
        '403':
          description: Constraint is violated
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/ReactionsPOSTResponse403'

  /configuration:
    post:
      description: Set new reactions configuration
      requestBody:
        description: |
          Reactions Storage Configuration.
          application/gzip: Gzip-encoded tar archive with 3 files (reaction.yaml, reaction_set.yaml, namespace.yaml)
        required: true
        content:
          appication/json:
            schema:
              $ref: 'definitions/requests.yaml#/components/schemas/ConfigurationPOSTRequest'
          application/gzip:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/ConfigurationPOSTResponse200'
        '415':
          description: Unsupported Content-Type
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/ConfigurationPOSTResponse415'
        '422':
          description: Configuration is not valid
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/ConfigurationPOSTResponse422'

  /configuration/namespace:
    get:
      description: Get namespace configuration
      parameters:
        - in: query
          name: namespace_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/NamespaceGETResponse200'
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/NamespaceGETResponse404'

  /configuration/available_reactions:
    get:
      description: Get available reactions for given namespace
      parameters:
        - in: query
          name: namespace_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/AvailableReactionsGETResponse200'
        '404':
          description: Namespace not found
          content:
            appliction/json:
              schema:
                $ref: 'definitions/responses.yaml#/components/schemas/AvailableReactionsGETResponse404'